Update the README and make the Vertex now uses doubles , added triangleTransform to apply transform on triangles and added the list vertexes to execute forEach loop on the vertexes of a triangle

+++ src/net/mako/rendering/Vertex.java
@@ -2,10 +2,10 @@ package net.mako.rendering;
 
 public class Vertex {
 
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
-    public Vertex(int x, int y){
+    public Vertex(double x, double y){
         this.x = x;
         this.y = y;
     }
